Test out the AxiLiteInitiator with Pynq
It works, but doesn't quite access memory correctly.

It's as though all reads (which are 32-bit) are forcibly 64-bit aligned.
So two successive 32-bit reads read the same thing, then the next jumps
up by 8 bytes.

This seems to be an issue with the AXI_HP0 port. If I switch it to the
AXI_GP0 port it works fine (with maybe a cycle more of latency?).

The AXI_HP0 port needs to be configured to act in 32-bit mode. It's in
64 by default. This fixes it:

devmem2 0xf8008000 w 1
devmem2 0xf8008014 w 1

And some sources say that this should only be done before the bitstream
is loaded, but it actually seems to work even while the bitstream is
running.