Implement Serial for ZynqGameboy
Note that this doesn't quite work with the TXS0108B. The serial output
when a cable isn't connected is correct, but when a cable is connected
the capacitance (presumably) throws things off and makes the outputs
oscillate and not be usable.

Directly connecting the Pynq-Z2 pins seems to work some of the time.
Presumably a better circuit would be better, maybe just open drain using
some transistors directly.