Rework clocking in ZynqGameboy to address timing issues
Previously, there were several asynchronous clocks:
* 100 MHz clock from PS, used to clock AXI target and initiator
  and as the input clock to the PLL:
* ~ 8 MHz clock used to make the gameboy clock
* The 25.2 and 126 MHz clocks used for HDMI output

There were lots of CDC issues between:
* AXI target config registers (100 MHz) and the Gameboy (8 MHz)
  (both in terms of readable properties, e.g. stat registers, and
  writable properties, like the 'running' flag)
* Gameboy emulated cartridge (8 Mhz) and the AXI initiator data source
  (100 MHz)

This commit switches how the clocks are generated and used. The 100 MHz
clock from the PS is now not used directly. Instead it's used as the
input to two PLLs: one that generates the 2 HDMI clocks, and one that
generates the 8MHz clock and a roughly 100MHz clock that's a precise
integer multiple of the other one.

The AXI Target (config registers) now runs at the same 8 MHz clock --
this wasn't high performance anyway, and it means it now runs in the
same clock domain as the gameboy, greatly simplifying things.

The emulated cartridge / AXI initiator runs at synchronous, integer
multiple of the Gameboy clock. There's still CDC to deal with, but it's
**greatly** simplified, and easier to show that it's correct.

The design passes timing, and even with heavy DRAM access on the PS
(which generally causes stalls and helps reveal instability), there are
no weird glitches that happen (indicating incorrect data was read from
the emulated ROM).

The newly re-added "cart stalls" stat register reveals that there are
more stalls than before -- roughly ~8000 per second than before. This is
still only 0.1%, imperceptible, but perhaps more stalls than strictly
needed are happening. Worth looking into, because this would be a bigger
problem with 2x GBC speed.